backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
spacy>=3.7.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
openai>=1.3.0
google-genai>=1.0.0
# google-generativeai>=0.8.0  # Deprecated - kept for fallback only
httpx>=0.25.0
Pillow>=10.0.0
pytesseract>=0.3.10
plotly>=5.18.0

